+++ its/src/test/java/com/sonar/it/shared/Tests.java
@@ -23,15 +23,11 @@ import com.sonar.orchestrator.Orchestrator;
 import com.sonar.orchestrator.build.BuildResult;
 import com.sonar.orchestrator.build.ScannerForMSBuild;
 import com.sonar.orchestrator.locator.MavenLocation;
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
-import org.apache.commons.io.FileUtils;
 import org.junit.ClassRule;
-import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
@@ -54,21 +50,12 @@ public class Tests {
     .addPlugin(MavenLocation.of("org.sonarsource.html", "sonar-html-plugin", "3.2.0.2082"))
     .build();
 
-  public static Path projectDir(TemporaryFolder temp, String projectName) throws IOException {
-    Path projectDir = Paths.get("projects").resolve(projectName);
-    FileUtils.deleteDirectory(new File(temp.getRoot(), projectName));
-    File newFolder = temp.newFolder(projectName);
-    Path tmpProjectDir = Paths.get(newFolder.getCanonicalPath());
-    FileUtils.copyDirectory(projectDir.toFile(), tmpProjectDir.toFile());
-    return tmpProjectDir;
+  static BuildResult analyzeProject(Path temp, String projectName) throws IOException {
+    return analyzeProject(temp, projectName, null);
   }
 
-  static BuildResult analyzeProject(TemporaryFolder temp, String projectName) throws IOException {
-    return analyzeProject(temp,projectName, null);
-  }
-
-  static BuildResult analyzeProject(TemporaryFolder temp, String projectName, @Nullable String profileKey, String... keyValues) throws IOException {
-    Path projectDir = Tests.projectDir(temp, projectName);
+  static BuildResult analyzeProject(Path temp, String projectName, @Nullable String profileKey, String... keyValues) throws IOException {
+    Path projectDir = TestUtils.projectDir(temp, projectName);
     ScannerForMSBuild beginStep = TestUtils.createBeginStep(projectName, projectDir)
       .setProfile(profileKey)
       .setProperties(keyValues);
